chore(EMS-1342): move custom resolvers into their own files
