+++ src/api/custom-queries/index.ts
@@ -1,0 +1,3 @@
+import getCompaniesHouseInformation from './get-companies-house-information';
+
+export { getCompaniesHouseInformation };
