+++ src/api/custom-resolvers/index.ts
@@ -6,6 +6,7 @@ import accountSignInSendNewCode from './account-sign-in-new-code';
 import verifyAccountSignInCode from './verify-account-sign-in-code';
 import addAndGetOTP from './add-and-get-OTP';
 import deleteApplicationByReferenceNumber from './delete-application-by-refrence-number';
+import updateExporterCompanyAndCompanyAddress from './update-exporter-company-and-company-address';
 import submitApplication from './submit-application';
 
 export {
@@ -17,5 +18,6 @@ export {
   verifyAccountSignInCode,
   addAndGetOTP,
   deleteApplicationByReferenceNumber,
+  updateExporterCompanyAndCompanyAddress,
   submitApplication,
 };
